requirements.txt
PyMySQL==0.7.10
opencv_python==3.3.0.10
numpy==1.12.1
tensorflow==1.4.0
Pillow==5.0.0
